Added contracts page and contracts link to NavBar. Added new Default page for Home screen. Changed table UI.
