+++ Bulk_Storage_Service/Bulk_Storage_Solutions/Content/Site.css
@@ -35,3 +35,7 @@ textarea {
     background-color: #777;
     border-color: #fff
 }
+
+.table-striped > tbody > tr:nth-child(2n+1) > td, .table-striped > tbody > tr:nth-child(2n+1) > th {
+    background-color: lightgray;
+}
