+++ Bulk_Storage_Service/Bulk_Storage_Solutions/Site.Master
@@ -51,6 +51,7 @@
                 <div class="navbar-collapse collapse">
                     <ul class="nav navbar-nav">
                         <li><a runat="server" href="~/Views/Default">Home</a></li>
+                        <li><a runat="server" href="~/Views/Contracts">Contracts</a></li>
                         <li><a runat="server" href="~/Views/Storage">Storage</a></li>
                         <li><a runat="server" href="~/Views/Contact">Contact</a></li>
                     </ul>
